% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmeth.R
\name{diffMeth}
\alias{diffMeth}
\title{Differential methylated CpGs detection}
\usage{
diffMeth(dat)
}
\arguments{
\item{dat}{GenomicRatioSet object}
}
\value{
DiffMeth object
}
\description{
Differential methylated CpGs detection
}
